# documentary/swagger-specs.yaml
openapi: 3.0.0
info:
  title: Drinks API
  version: 1.0.0
  description: API for managing drinks
  servers:
  - url: http://localhost:3000/

paths:
  /drinks:
    get:
      summary: Get all drinks
      description: Retrieve a list of all drinks from the database.
      responses:
        '200':
          description: A list of drinks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Drink'
        '500':
          description: Internal server error.
    post:
      summary: Create a new drink
      description: Create a new drink in the database.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Drink'
      responses:
        '201':
          description: Drink created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  drinkId:
                    type: integer
                    description: The ID of the newly created drink.
        '409':
          description: A drink with that name already exists.
        '500':
          description: Internal server error.
  /drinks/{id}:
    get:
      summary: Get a drink by ID
      description: Retrieve a specific drink by its ID.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: The ID of the drink.
      responses:
        '200':
          description: The drink with the specified ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drink'
        '404':
          description: Drink not found.
        '500':
          description: drink creation unsuccessful.
    put:
      summary: Update a drink
      description: Update an existing drink in the database.
      parameters:
        - in: path
          name: drinkId
          schema:
            type: integer
          required: true
          description: The ID of the drink to update.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Drink'
      responses:
        '200':
          description: Drink updated successfully.
        '404':
          description: Drink not found.
        '500':
          description: Drink not updated.
    delete:
      summary: Delete a drink
      description: Delete a drink from the database.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: The ID of the drink to delete.
      responses:
        '200':
          description: Drink deleted successfully.
        '404':
          description: Drink not found.
        '500':
          description: Internal server error.
components:
  schemas:
    Drink:
      type: object
      properties:
        id:
          type: integer
          description: The ID of the drink.
        name:
          type: string
          description: The name of the drink.
        description:
          type: string
          description: A description of the drink.
        alcoholic:
          type: boolean
          description: Whether the drink is alcoholic.
        category:
          type: string
          description: The category of the drink.
        image_path:
          type: string
          description: The URL of the drink's image.
        instructions:
          type: string
          description: Instructions for making the drink.
        ingredients:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                description: The name of the ingredient.
              quantity:
                type: number
                description: The quantity of the ingredient.
              measurement:
                type: string
                description: The measurement unit for the ingredient.